fix: update tsconfig issue with vscode. Extend child configuration file
**npm run build** is not working due to the use of `tsconfigRootDir` on the eslint config file

#Cause
vscode was not recognizing the tsconfig if it not there one file in the root. We updated each eslint to inform where was each tsconfig using the `tsconfigRootDir` and vscode back to work but npm run build breaks.

# Fix
use a tsconfig in the root directory and extend each folder specific tsconfig with only redifined values like srcDir.

![image.png](https://dev.azure.com/711digital/1dcb64b4-79dc-4bec-9438-d7c2585ab9c0/_apis/git/repositories/8ba7673d-db29-45cf-9b73-322c0ae76812/pullRequests/10/attachments/image%20(2).png)

Error from VsCode after set `tsconfigRootDir`
![image.png](https://dev.azure.com/711digital/1dcb64b4-79dc-4bec-9438-d7c2585ab9c0/_apis/git/repositories/8ba7673d-db29-45cf-9b73-322c0ae76812/pullRequests/10/attachments/image.png)

Related work items: #345