feat: Add S3 cloud storage interface implementation
This S3 adapter is an implementation of the cloudStorage interface and connect with amazonstorage service in order to read files from a specified bucket, the AWS-SDK method used for read data is getObject which is a method from the AWS.S3 client. AWS returns an object that contains the bucket name, file name, metadata, body etc. We focus on get the body content that is an hexadecimal buffer that contains the data we require and we are returning it as a promise.

MC#174