Merged PR 244: fix: Default azure response body to null
By default if the azure response body is not set (aka `undefined`), then the azure functions runtime will incorrectly assume the full response object is actually the response body.  By changing the default value of body to `null` we are able to work around this awkward behavior.

- [x] Defaults response body to `null` if not defined
- [x] Updated unit tests

Related work items: #999