Merged PR 242: fix: Update TestModule condition
In order to fix the issues in the @multicloud/sls-azure and
@multicloud/sls-aws packages that occur after the TestModule
implementation, its required to change the condition that decides
whether to add the TestModule or not by using an AND operator instead
of the OR operator.
Meaning that the TestModule will be only added to the modules chain if
there are no modules and the environment is test.

Related work items: #1001