feat: Exposes `context.error` method to signal an error has occurred during middleware pipeline (#10)
Adds the ability to manually call into the ExceptionMiddleware via a new context.error(...) method to set the error an optionally a status code.

Closes N/A

When working with callbacks (specifically async callbacks) throwing an error in an async handler cannot be handled by a typical try/catch block. In this case we need a way to mark that an exception has occurred without having an unhanded exception silently bubble up the call stack and terminate the process.