fix: Update the S3Storage to work with both node version 8.x and 10.x
We modified the S3Storage file in order to read the readable object length both in node version 8.x and 10.x. We did that because when deployed an AWS handler that uploads a file in a bucket, the readable object in the 8.x node version doesn't have the readableLength property.