feat: Update the sls-awslibrary to support the deprecated method
We added the old invoke method to the LambdaCloudService as a deprecated method to don't modify the CloudService consumers. We overloaded the invoke method and took the corresponding action in the implementation method if is called the deprecated or the new method. This new feature only works when the sls-core library is updated with the CloudService latest changes.