feat: We implemented the CloudRequest's path property in the AWSRequest class
We added the path property in the AWSRequest class and fill it in the constructor class taking the value from the request event URL. We also added one unit test to verify the property is filled with the correct value. We need to update the sls-core version when it's updated with the path's property change.